New open source particle trajectory integrator. Switched from Numerical Recipe's stepperdopr853 to Boost::Odeint's runge_kutta_dopri5 integrator for particle trajectory integration. Replaced most of the unnecessary long double calculations with double.

+++ bruteforce.cpp
@@ -62,8 +62,8 @@ void TBFIntegrator::operator()(const state_type &y, value_type x){
 			<< B[0]/BFBws << " " << B[1]/BFBws << " " << B[2]/BFBws << '\n';
 }
 
-long double TBFIntegrator::Integrate(double x1, long double y1[6], long double B1[4][4],
-					double x2, long double y2[6], long double B2[4][4], std::ofstream *&spinout){
+long double TBFIntegrator::Integrate(double x1, double y1[6], long double B1[4][4],
+					double x2, double y2[6], long double B2[4][4], std::ofstream *&spinout){
 	if (gamma == 0)
 		return 1;
